% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdr.R
\name{calc_fdr_col}
\alias{calc_fdr_col}
\title{Calculate the empirical false discovery rate (FDR).}
\usage{
calc_fdr_col(post_prob, threshold, col_name = NULL, col_idx = NULL)
}
\arguments{
\item{post_prob}{A numeric matrix of posterior probabilities.}

\item{threshold}{Numerical value of the posterior probability threshold at which to
calculate the empirical FDR.}

\item{col_name}{Character string of the column name (optional).}

\item{col_idx}{Integer of the column index (optional).}
}
\value{
Numerical value of the empirical false discovery rate (FDR).
}
\description{
Calculate the empirical false discovery rate (FDR) of a
column in a matrix of posterior probabilities given a specified threshold.
}
\details{
At least one of \code{col_name} or \code{col_idx} must be specified.
}
