% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdr.R
\name{calc_fdr_finemap}
\alias{calc_fdr_finemap}
\title{Calculate the empirical false discovery rate (FDR) after fine-mapping.}
\usage{
calc_fdr_finemap(post_prob, threshold, fail_idx)
}
\arguments{
\item{post_prob}{numeric vector of posterior probabilities.}

\item{threshold}{numerical value of the posterior probability threshold at which to
calculate the empirical FDR.}

\item{fail_idx}{integer vector of the indices of observations which failed
fine-mapping. These are the observations where the association pattern with the
highest posterior probability after fine-mapping changed (if using collapsed
categories, they should have changed such they are no longer in an
association pattern fitting the description of the collapsed category).}
}
\value{
A numeric value of the empirical false discovery rate (FDR).
}
\description{
Calculate the empirical false discovery rate (FDR) given
a vector of posterior probabilities and a specified threshold,
adjusting for false positives identified by fine-mapping.
}
