% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdr.R
\name{calc_fdr_multi}
\alias{calc_fdr_multi}
\title{Calculate the empirical false discovery rate (FDR).}
\usage{
calc_fdr_multi(post_prob, threshold, col_names = NULL, col_idxs = NULL)
}
\arguments{
\item{post_prob}{A numeric matrix of posterior probabilities.}

\item{threshold}{Numerical value of the posterior probability threshold at which to
calculate the empirical FDR.}

\item{col_names}{A character vector of column names (optional).}

\item{col_idxs}{An integer vector of column indices (optional).}
}
\value{
Numerical value of the empirical false discovery rate (FDR).
}
\description{
Calculate the empirical false discovery rate (FDR) of specified
columns in a matrix of posterior probabilities given a specified threshold.
}
\details{
At least one of \code{col_names} or \code{col_idxs} must be specified.
}
