% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{collapse_pp_combin}
\alias{collapse_pp_combin}
\title{Collapse posterior probabilities of association for combinations of individual traits.}
\usage{
collapse_pp_combin(post_prob, combos, req_idx = NULL, prefix = "pp")
}
\arguments{
\item{post_prob}{matrix of posterior probabilities.}

\item{combos}{matrix or integer specifying combinations. When matrix, each row is
a combination, with column indices of traits belonging to the combination specified in
columns of \code{combos}. When integer, all combinations of (total # traits)
taking \code{combos} elements at a time will be tested.}

\item{req_idx}{(optional) scalar or integer vector of trait(s) where a non-null
association is required (i.e. from the alternative distribution).
If \code{NULL}, no traits will be required.}

\item{prefix}{character string denoting prefix of column names for the results matrix.}
}
\value{
A numeric matrix of combined posterior probabilities.
}
\description{
Combine the posterior probabilities for combinations of individual traits/studies according to the
association patterns from which those traits/studies all come from the alternative distribution.
Provides the posterior probability of being associated with "at least traits
\eqn{X}, \eqn{Y} and \eqn{Z}", for example.
Can also require non-null association with one or more traits/studies and combine
over patterns where that/those trait(s) also come from the alternative distribution.
}
\examples{
## all pairs of traits
myCombos <- t(combn(4,2))
collapsed_pp <- collapse_pp_combin(post_prob,myCombos)

}
