% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{collapse_pp_num}
\alias{collapse_pp_num}
\title{Collapse posterior probabilities based on number of traits of association.}
\usage{
collapse_pp_num(post_prob, req_idx = NULL, prefix = "pp_ge_")
}
\arguments{
\item{post_prob}{matrix of posterior probabilities.}

\item{req_idx}{(optional) scalar or integer vector of trait(s) where a non-null
association is required (i.e. from the alternative distribution).
If \code{NULL}, no traits will be required.}

\item{prefix}{character string denoting prefix of column names for the results matrix.}
}
\value{
A numeric matrix of combined posterior probabilities.
}
\description{
Combine the posterior probabilities of association patterns
according to the number of traits with non-null associations in each pattern.
Provides the posterior probability of
being associated with "at least \eqn{n}" number of traits/studies
by summing over association patterns with at least \eqn{n} studies
coming from the alternative distribution. Can also require non-null
association with one or more traits and combine over other traits.
}
