% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{e_step}
\alias{e_step}
\title{E-step}
\usage{
e_step(old_pi, Dmat)
}
\arguments{
\item{old_pi}{vector of association pattern proportions, fit through maximization.}

\item{Dmat}{matrix of conditional joint densities under each association pattern.}
}
\value{
A matrix estimating the posterior expectations
(i.e. estimated proportion of observations belonging to each association pattern).
}
\description{
Calculate posterior expectations given maximized estimate for \eqn{\pi}, the
proportion of observations (e.g. SNPs) coming from each association pattern.
}
