% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{e_step_Dmat}
\alias{e_step_Dmat}
\title{E Step, using precalculated joint densities}
\usage{
e_step_Dmat(old_pi, Dmat)
}
\arguments{
\item{old_pi}{vector of configuration proportions, fit through maximization
(usualy 2^J for J data types)}

\item{Dmat}{matrix of conditional joint densities under each configuration}
}
\value{
Returns a matrix estimating the posterior expectations
(i.e. estimated probability of each configuration for each SNP).
}
\description{
Calculate posterior expectations given maximized estimate for \eqn{\pi} (the
proportion of SNPs coming from each configuration).
}
