% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{e_step_withColSums}
\alias{e_step_withColSums}
\title{E Step, using precalculated joint densities}
\usage{
e_step_withColSums(old_pi, Dmat)
}
\arguments{
\item{old_pi}{vector of configuration proportions, fit through maximization
(usualy 2^J for J data types)}

\item{Dmat}{matrix of conditional joint densities under each configuration}
}
\value{
Returns a matrix of column sums of posterior expectations
(allows processing to be performed in chunks).
}
\description{
Calculate posterior expectations given maximized estimate for \eqn{\pi} (the
proportion of SNPs coming from each configuration).
}
