% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{e_stepC}
\alias{e_stepC}
\title{E Step}
\usage{
e_stepC(old_pi, Q, D_0, D_1)
}
\arguments{
\item{old_pi}{vector of configuration proportions, fit through maximization
(usualy 2^J for J data types)}

\item{Q}{matrix of configurations}

\item{D_0}{estimate for the null density function}

\item{D_1}{estimate for the alternative density function}
}
\value{
Returns a vector estimating the posterior expectations
(i.e. estimated probability of each configuration for each SNP).
}
\description{
Calculate posterior expectations given maximized estimate for \eqn{\pi} (the
proportion of SNPs coming from each configuration).
}
