% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_stepsR.R
\name{e_stepR}
\alias{e_stepR}
\title{E Step}
\usage{
e_stepR(old_pi, n_obs, n_pattern, q, density_0, density_1, cl = NULL)
}
\arguments{
\item{old_pi}{vector of configuration proportions, fit through maximization}

\item{n_obs}{numerical value; number of observations (e.g. SNPs/loci)}

\item{n_pattern}{numerical value; number of possible patterns
(usualy 2^J for J data types)}

\item{q}{matrix of configurations}

\item{density_0}{estimate for the null density function}

\item{density_1}{estimate for the alternative density function}

\item{cl}{cluster object (optional)}
}
\value{
Returns a vector estimating the posterior expectations
(i.e. estimated probability of each configuration for each SNP).
}
\description{
DEPRECATED (replaced by faster RcppArmadillo version).
Calculate posterior expectations given maximized estimate for \eqn{\pi} (the
proportion of SNPs coming from each configuration).
}
