% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdr.R
\name{efdr}
\alias{efdr}
\title{Calculate empirical False Discovery Rate (eFDR)}
\usage{
efdr(true_res, perm_res, thresholds = seq(0.99999, 0.999, -1e-05),
  config = NULL)
}
\arguments{
\item{true_res}{a list; results returned by \code{\link{estimate_config}} run
with real data}

\item{perm_res}{a list of lists; results returned by \code{\link{estimate_config}} run
with permuted datasets}

\item{thresholds}{vector of posterior probability thresholds for which to calculate eFDR}

\item{config}{numerical value; the column number of the configuration for which to
estimate eFDR values}
}
\value{
vector of the eFDR for each posterior probability threshold tested.
}
\description{
Calculate the empirical False Discovery Rate (eFDR) given
observed results and results from permuted datasets. Takes
a vector of potential posterior probability thresholds and
calculates the eFDR for each threshold.
}
