% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrative_func.R
\name{estimate_config}
\alias{estimate_config}
\title{Estimate Posterior Probabilities of Configurations}
\usage{
estimate_config(betas, sds, mafs, dfs, priors, tol = 0.001, par_size = 0)
}
\arguments{
\item{betas}{matrix of coefficient estimates.}

\item{sds}{matrix of standard errors (for coefficient estimates).}

\item{mafs}{vector of minor allele frequencies (MAFs).}

\item{dfs}{vector of degrees of freedom.}

\item{priors}{vector of prior (marginal) probabilities.}

\item{tol}{numerical value; specifies tolerance threshold for convergence.}

\item{par_size}{numerical value; specifies the number of CPUs/cores/processors for
parallel computing
(0 for sequential processing).}
}
\value{
A list with the following elements:
\tabular{ll}{
\code{post_prob} \tab matrix of posterior probabilities
(rows are SNPs; columns are configurations)\cr
\code{config_prop} \tab vector of estimated proportion of SNPs
belonging to each configuration\cr
\code{Tstat_m} \tab matrix of moderated t-statistics\cr
\code{D0} \tab matrix of densities calculated under the null distribution\cr
\code{D1} \tab matrix of densities calculated under the alternative distribution\cr
\code{priors} \tab vector of prior probabilities used in the estimation\cr
\code{tol} \tab numerical value; the tolerance threshold used in determining convergence
}
}
\description{
For each SNP, estimates the posterior probability for each configuration.
Utilizes parallel computing, when available.
}
\details{
The following are additional details describing the input arguments:
\tabular{ll}{
\code{betas} \tab  coefficients should be an LxD matrix: L = number of SNPs ("Loci");
D = number data sources.\cr
\code{sds} \tab standard errors should be an LxD matrix: L = number of SNPs ("Loci");
D = number data sources.\cr
\code{mafs} \tab minor allele frequencies (MAF) should be vector with length matching
the number of SNPs/Loci\cr
 \tab (same order as rows in \code{betas} and \code{sds}).\cr
\code{dfs} \tab vector containing number of subjects in each dataset\cr
 \tab (same order as columns in \code{betas} and \code{sds})\cr
\code{priors} \tab marginal (prior) probability that a test-statistic comes from
the alternative distribution for each data source\cr
 \tab (same order as columns in \code{betas} and \code{sds})
}
}
