% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density_estimation.R
\name{estimate_densities_modT}
\alias{estimate_densities_modT}
\title{Estimate parameters under the null and alternative densities}
\usage{
estimate_densities_modT(betas, sds, df, alt_prop, mafs = NULL,
  N = NULL)
}
\arguments{
\item{betas}{vector of coefficient estimates.}

\item{sds}{vector of standard errors (for coefficient estimates).}

\item{df}{scalar or vector of first degrees of freedom of the F-distribution
(usually, \eqn{N-p-1}).}

\item{alt_prop}{proportion of test statistics from the alternative density.}

\item{mafs}{vector of minor allele frequencies (MAFs) (optional).
If \code{NULL}, no adjustment will be made for MAF in variance calculations.}

\item{N}{scalar or vector of number of subjects (optional).
Should be specified if \code{mafs!=NULL}.}
}
\value{
A list with the following elements:
\tabular{ll}{
\code{Tstat_mod} \tab matrix of moderated t-statistics.\cr
\code{df_mod} \tab vector of the moderated degrees of freedom.\cr
\code{scaler} \tab vector of the scaling factors for the moderated t-statistics under the alternative.\cr
\code{prior_df} \tab scalar of the prior degrees of freedom.\cr
\code{prior_var} \tab scalar of the prior variance estimator.\cr
\code{unscaled_var} \tab scalar of the unscaled variance prior on non-zero coefficients.
}
}
\description{
Estimate the parameters of the null and alternative
density functions using moderated \eqn{t}-statistics.
}
\details{
Following Smyth (2004), the function calculates moderated \eqn{t}-statistics using
conditional posterior means of the variance. The moderated \eqn{t}-statistics are used to estimate
the null density under a \eqn{t}-distribution and
the alternative density under a scaled \eqn{t}-distribution, each with moderated degrees of freedom.
}
