% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{find_leadSNPs}
\alias{find_leadSNPs}
\title{Find the lead SNP for each phenotype in each region.}
\usage{
find_leadSNPs(data, SNP_col, pheno_cols, stat_cols, data_type = "pvalue",
  suffices = NULL)
}
\arguments{
\item{data}{A data.table. Each row will be a SNP-phenotype combination
with statistics necessary to determine the lead SNP in each phenotype region.}

\item{SNP_col}{Character string of the column name of the SNP (must be "SNP"
for current version).}

\item{pheno_cols}{Character vector of the column names of the phenotypes.}

\item{stat_cols}{Character vector of the column names of statistics to be
used to determine lead SNPs.}

\item{data_type}{Character string denoting type of statistics being used. Must be
either "pvalue" or "tstat".}

\item{suffices}{A character vector denoting suffices to use for the names of the
lead SNP columns (optional). If \code{NULL}, consecutive integers will be assigned.}
}
\value{
A data.table containing information about the lead SNPs and
associated statistics. The columns will be \code{pheno_cols} followed by two columns
for each phenotype: the name of the lead SNP and the value of the statistic for
that lead SNP in that phenotype.
}
\description{
Subset results from Primo output based on a vector of indices.
}
