% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finemap.R
\name{fine_map_once}
\alias{fine_map_once}
\title{Setup fine-mapping for a specified variant by index.}
\usage{
fine_map_once(Primo_res, IDs, idx, leadSNPs_byRegion, SNP_col, pheno_cols,
  snp.info, LDmat, LD_thresh = 1, dist_thresh = 0, pval_thresh = 1,
  suffices = 1:length(pheno_cols))
}
\description{
Setup fine-mapping for a specified variant by index.
}
