% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{m_stepC}
\alias{m_stepC}
\title{M Step}
\usage{
m_stepC(old_B)
}
\arguments{
\item{old_B}{matrix of posterior expectations/probabilities ; usually
from an expectation (E-)step}
}
\value{
Returns a vector estimating the proportion of SNPs coming from each configuration.
}
\description{
Estimates the \eqn{\pi} vector that maximizes the posterior expectation function from the E-step.
}
