% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_stepsR.R
\name{m_stepR}
\alias{m_stepR}
\title{M Step}
\usage{
m_stepR(old_b, n_obs, n_pattern)
}
\arguments{
\item{old_b}{matrix of posterior expectations/probabilities ; usually
from an expectation (E-)step}

\item{n_obs}{numerical value; number of observations (e.g. SNPs/loci)}

\item{n_pattern}{numerical value; number of possible patterns
(usually \eqn{2^J} for J data types)}
}
\value{
Returns a vector estimating the proportion of SNPs coming from each configuration.
}
\description{
DEPRECATED (replaced by faster RcppArmadillo version).
Estimates the \eqn{\pi} vector that maximizes the posterior expectation function from the E-step.
}
