% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permute_func.R
\name{permute_integ}
\alias{permute_integ}
\title{Estimate Posterior Probabilities for Permuted Datasets}
\usage{
permute_integ(betas, sds, mafs, dfs, alt_proportions, tol = 0.001,
  par_size = 0, perm_par_size = 0)
}
\arguments{
\item{betas}{matrix of coefficient estimates.}

\item{sds}{matrix of standard errors (for coefficient estimates).}

\item{mafs}{vector of minor allele frequencies (MAFs).}

\item{dfs}{vector of degrees of freedom.}

\item{alt_proportions}{vector of the proportions of test-statistics used in estimating
alternative densities.}

\item{tol}{numerical value; specifies tolerance threshold for convergence.}

\item{par_size}{numerical value; specifies the number of CPUs/cores/processors for
parallel computing
(0 for sequential processing).}

\item{perm_par_size}{numerical value; specifies the number of CPUs/cores/processors for
parallel computing of permutations(0 for sequential processing).}
}
\value{
A list of lists, where each list holds results from a single run of
\code{\link{estimate_config}}. Each run represents results using one permuted dataset.
See \code{\link{estimate_config}} for details of each list.
}
\description{
Permutes several datasets (one column permuted in each)
and estimates the posterior probability for each configuration
for each SNP under each permuted dataset.
Utilizes parallel computing, when available.
}
\details{
\code{perm_par_size} differs from \code{par_size} in that permuted datasets
are processed in parallel, rather than configurations. Developer recommendation is
to parallelize permutations, running configurations sequentially.
See documentation for \code{\link{estimate_config}} for additional details
regarding the input arguments.
}
