% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permute_func.R
\name{permute_multi}
\alias{permute_multi}
\title{Estimate Posterior Probabilities for Multiple Permuted Datasets}
\usage{
permute_multi(betas, sds, mafs, dfs, true_res, nperm = 10, par_size = 0,
  perm_par_size = 0, density_list = NULL)
}
\arguments{
\item{betas}{matrix of coefficient estimates.}

\item{sds}{matrix of standard errors (for coefficient estimates).}

\item{mafs}{vector of minor allele frequencies (MAFs).}

\item{dfs}{vector of degrees of freedom.}

\item{true_res}{a list; results returned by \code{\link{estimate_config}} when
run with true data}

\item{nperm}{numerical value; number of permutations to run}

\item{par_size}{numerical value; specifies the number of CPUs/cores/processors for
parallel computing (0 for sequential processing).}

\item{perm_par_size}{numerical value; specifies the number of CPUs/cores/processors for
parallel computing of permutations(0 for sequential processing).}

\item{density_list}{(optional) list of densities estimated by \code{estimate_densities()}.}
}
\value{
A list of lists, where each list holds the matrices of posterior probabilities
from \code{\link{estimate_config}} run on a permuted dataset (i.e. a single column).
}
\description{
Run multiple permutations of multi-omics datasets.
Each set of permutations permutes one column at a time
and estimates the posterior probability for each configuration
for each SNP under each permuted dataset.
}
\details{
See documentation for \code{\link{estimate_config}} for additional details
regarding the input arguments.
}
