% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permute_func.R
\name{permute_once_dens}
\alias{permute_once_dens}
\title{Estimate Posterior Probabilities after Permuting Densities}
\usage{
permute_once_dens(alt_props, perm_col, tol = 0.001, density_list, config_prop)
}
\arguments{
\item{alt_props}{vector of the proportions of test-statistics used in estimating
alternative densities.}

\item{perm_col}{(vector of) numerical value(s); column number(s) to permute}

\item{tol}{numerical value; specifies tolerance threshold for convergence.}

\item{density_list}{list of densities estimated by \code{estimate_densities()}.}

\item{config_prop}{vector of estimated proportion of SNPs
belonging to each configuration}
}
\value{
A list with the following elements, based on permuted data:
\tabular{ll}{
\code{post_prob} \tab matrix of posterior probabilities
(rows are SNPs; columns are configurations)\cr
\code{config_prop} \tab vector of estimated proportion of SNPs
belonging to each configuration\cr
\code{Tstat_mod} \tab matrix of moderated t-statistics\cr
\code{D0} \tab matrix of densities calculated under the null distribution\cr
\code{D1} \tab matrix of densities calculated under the alternative distribution\cr
\code{alt_props} \tab vector of the proportions of test-statistics used in
estimating alternative densities\cr
\code{tol} \tab numerical value; the tolerance threshold used in determining convergence
}
}
\description{
Permute densities for one or more columns and estimate the posterior probability
for each configuration for each SNP under the permuted dataset.
}
\details{
See documentation for \code{\link{estimate_config}} for additional details
regarding the input arguments.
}
