% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permute_func.R
\name{permute_once_stats}
\alias{permute_once_stats}
\title{Estimate Posterior Probabilities after Permuting One Column of Statistics}
\usage{
permute_once_stats(betas, sds, mafs, dfs, alt_props, perm_col, tol = 0.001,
  par_size = 0)
}
\arguments{
\item{betas}{matrix of coefficient estimates.}

\item{sds}{matrix of standard errors (for coefficient estimates).}

\item{mafs}{vector of minor allele frequencies (MAFs).}

\item{dfs}{vector of degrees of freedom.}

\item{alt_props}{vector of the proportions of test-statistics used in estimating
alternative densities.}

\item{perm_col}{numerical value; column number to permute}

\item{tol}{numerical value; specifies tolerance threshold for convergence.}

\item{par_size}{numerical value; specifies the number of CPUs/cores/processors for parallel computing
(0 for sequential processing).}

\item{density_list}{(optional) list of densities estimated by \code{estimate_densities()}.}

\item{perm_densities}{logical value; when true, permutes the previously calculated densities
instead of reestimating densities from permuted betas and sds}
}
\value{
A list with the following elements, based on permuted data:
\tabular{ll}{
\code{post_prob} \tab matrix of posterior probabilities
(rows are SNPs; columns are configurations)\cr
\code{config_prop} \tab vector of estimated proportion of SNPs
belonging to each configuration\cr
\code{Tstat_mod} \tab matrix of moderated t-statistics\cr
\code{D0} \tab matrix of densities calculated under the null distribution\cr
\code{D1} \tab matrix of densities calculated under the alternative distribution\cr
\code{alt_props} \tab vector of the proportions of test-statistics used in
estimating alternative densities\cr
\code{tol} \tab numerical value; the tolerance threshold used in determining convergence
}
}
\description{
Permute one column (beta and sd) and estimate the posterior probability for each configuration
for each SNP under the permuted dataset.
}
\details{
See documentation for \code{\link{estimate_config}} for additional details
regarding the input arguments.
}
