% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primo.R
\docType{package}
\name{primo-package}
\alias{primo}
\alias{primo-package}
\title{primo: Package in R for Integrating Multi-Omics data}
\description{
Implements methods to integrate multi-omics data to detect joint associations 
              and pleiotropy effects across more than 3 eQTL and GWAS studies. 
              It can also be used more broadly to infer multi-way associations 
              using omics data from more than 2 data types or sources, beyond eQTL and GWAS.
}
\details{
To estimate the posterior probabilities of belonging to each association pattern
(i.e. configuration), use the \code{\link{estimate_config}} function.
}
\author{
\strong{Maintainer}: Kevin J. Gleason \email{kevin.j.gleason@gmail.com}

Authors:
\itemize{
  \item Fan Yang \email{FAN.3.YANG@ucdenver.edu}
  \item Lin S. Chen \email{lchen@health.bsd.uchicago.edu}
}

}
