% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{Primo}
\alias{Primo}
\title{Estimate posterior probabilities of association patterns.}
\usage{
Primo(betas = NULL, sds = NULL, dfs = NULL, pvals = NULL,
  alt_props, mafs = NULL, N = NULL, Gamma = NULL, tol = 0.001,
  par_size = 1, use_method = "tstat")
}
\arguments{
\item{betas}{matrix of coefficient estimates.}

\item{sds}{matrix of standard errors (for coefficient estimates).}

\item{dfs}{vector or matrix of degrees of freedom.}

\item{pvals}{matrix of \eqn{P}-values.}

\item{alt_props}{vector of the proportions of test statistics from the
alternative densities.}

\item{mafs}{vector or matrix of minor allele frequencies (MAFs).}

\item{N}{vector or matrix of number of subjects.}

\item{Gamma}{correlation matrix.}

\item{tol}{numeric value specifying the tolerance threshold for convergence.}

\item{par_size}{numeric value specifying the number of workers for
parallel computing (1 for sequential processing).}

\item{use_method}{character string denoting which method to use.
Must be one of "tstat" or "pval".}
}
\value{
A list with the following elements:
\tabular{ll}{
\code{post_prob} \tab matrix of posterior probabilities
(each column corresponds to an association pattern).\cr
\code{pis} \tab vector of estimated proportion of observations
belonging to each association pattern.\cr
\code{D_mat} \tab matrix of densities under each association pattern.\cr
\code{Gamma} \tab correlation matrix.\cr
}

\itemize{
\item If \code{use_method}="tstat", the list will additionally contain:
\tabular{ll}{
\code{Tstat_mod} \tab matrix of moderated t-statistics.\cr
\code{V_mat} \tab matrix of scaling factors under the alternative distribution.\cr
\code{mdf_sd_mat} \tab matrix of standard deviation adjustment according to
 moderated degrees of freedom: \code{sqrt(df/(df-2))}.\cr
\code{prior_df} \tab vector of the prior degrees of freedom for each marginal distribution.\cr
\code{prior_var} \tab vector of the prior variance estimators for each marginal distribution.\cr
\code{unscaled_var} \tab vector of the unscaled variance priors on non-zero coefficients
for each marginal distribution.
}

\item If \code{use_method}="pval", the list will additionally contain:
\tabular{ll}{
\code{chi_mix} \tab matrix of \eqn{-2}log(\eqn{P})-values.\cr
\code{A} \tab vector of scaling factors under the alternative distributions.\cr
\code{df_alt} \tab vector of degrees of freedom approximated for the alternative distributions.\cr
 }
 }

The primary element of interest for inference is the posterior probabilities matrix, \code{post_prob}.
The estimated proportion of observations belonging to each association pattern, \code{pis}, may
also be of interest.
The remaining elements are returned primarily for use by other functions.
}
\description{
For each observation (e.g. SNP), estimate the posterior probability for each association pattern.
This function calls either the \eqn{t}-statistic (\code{\link{Primo_tstat}}) or
\eqn{P}-value (\code{\link{Primo_pval}}) version of Primo.
Utilizes parallel computing, when available. For package documentation, see \code{\link{Primo-package}}.
}
\details{
The following are additional details describing the input arguments
 (for \eqn{m} SNPs/observations measured in \eqn{d} studies):

\itemize{

\item \code{alt_props} (vector of length \eqn{d}) must be specified whether
\code{use_method}="tstat" or \code{use_method}="pval".

\item If \code{use_method}="tstat", the following arguments must be specified (not \code{NULL}):
  \tabular{ll}{
    \code{betas} \tab  \eqn{m} x \eqn{d} matrix.\cr
    \code{sds} \tab \eqn{m} x \eqn{d} matrix.\cr
    \code{dfs} \tab vector of length \eqn{d} or an \eqn{m} x \eqn{d} matrix.\cr
  }

  If the observations are SNPs, \code{mafs} and \code{N} can optionally be specified in order to adjust
  error variances for minor allele frequencies.
  \code{mafs} should be either a vector of length \eqn{m} or an \eqn{m} x \eqn{d} matrix.
  \code{N} should be either a vector of length \eqn{m} or an \eqn{m} x \eqn{d} matrix.

\item If \code{use_method}="pval", the following arguments must be specified (not \code{NULL}):
  \tabular{ll}{
    \code{pvals} \tab  \eqn{m} x \eqn{d} matrix.\cr
  }
}
}
