% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{Primo_chiMix}
\alias{Primo_chiMix}
\title{Estimate posterior probabilities of association patterns, using mixture of chi-squared statistics.}
\usage{
Primo_chiMix(chi_mix, A, df_alt, Gamma, tol = 0.001, par_size = 1)
}
\arguments{
\item{chi_mix}{matrix of \eqn{-2}log(\eqn{P})-values from test statistics.}

\item{A}{vector of scaling factors under the alternative distributions.\cr}

\item{df_alt}{vector of degrees of freedom approximated for the alternative distributions.\cr}

\item{Gamma}{correlation matrix.}

\item{tol}{numeric value specifying tolerance threshold for convergence.}

\item{par_size}{numeric value specifying the number of workers for
parallel computing (1 for sequential processing).}
}
\value{
A list with the following elements:
\tabular{ll}{
\code{post_prob} \tab matrix of posterior probabilities
(each column corresponds to an association pattern).\cr
\code{pis} \tab vector of estimated proportion of observations
belonging to each association pattern.\cr
\code{D_mat} \tab matrix of densities under each association pattern.\cr
\code{Gamma} \tab correlation matrix.\cr
\code{chi_mix} \tab matrix of \eqn{-2}log(\eqn{P})-values.\cr
\code{A} \tab vector of scaling factors under the alternative distributions.\cr
\code{df_alt} \tab vector of degrees of freedom approximated for the alternative distributions.\cr
}

The main element of interest for inference is the posterior probabilities matrix, \code{post_prob}.
The estimated proportion of observations belonging to each association pattern, \code{pis}, may
also be of interest. The remaining elements are returned primarily for use by other functions.
}
\description{
For each observation (e.g. SNP), estimates the posterior probability of
each association pattern. This version of the function uses a mixture of chi-squared statistics
and the parameters of the alternative distributions which were previously calculated
(i.e. using \code{\link{estimate_densities_pval}}). It is useful for cases where
the same statistic from one study (e.g. gene-SNP pair) may be mapped to
multiple statistics from another study (e.g. multiple gene-CpG pairings).
Utilizes parallel computing, when available.
}
\details{
The following are additional details describing the input arguments
 (for \eqn{m} SNPs/observations measured in \eqn{d} studies):
\tabular{ll}{
\code{chi_mix} \tab  \eqn{m} x \eqn{d} matrix.\cr
\code{A} \tab vector of length \eqn{d}.\cr
\code{df_alt} \tab vector of length \eqn{d}.\cr
\code{Gamma} \tab  \eqn{d} x \eqn{d} matrix.
}
}
