% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{Primo_missData_pval}
\alias{Primo_missData_pval}
\title{Estimate posterior probabilities for observations missing from original Primo analysis.}
\usage{
Primo_missData_pval(pvals, trait_idx, pis, Gamma, A, df_alt,
  par_size = 1)
}
\arguments{
\item{pvals}{matrix of \eqn{P}-values from test statistics.}

\item{trait_idx}{integer vector of the columns corresponding to non-missing phenotypes/studies.}

\item{pis}{matrix (one-row) of the estimated proportion of observations
belonging to each association pattern}

\item{Gamma}{correlation matrix.}

\item{A}{vector of scaling factors under the alternative distributions.}

\item{df_alt}{vector of degrees of freedom approximated for the alternative distributions.}

\item{par_size}{numeric value; specifies the number of CPUs/cores/processors for
parallel computing (1 for sequential processing).}

\item{chi_mix}{matrix of \eqn{-2}log(\eqn{P})-values.}
}
\value{
A list with the following elements:
\tabular{ll}{
\code{post_prob} \tab matrix of posterior probabilities
(rows are SNPs; columns are association patterns).\cr
\code{pis} \tab vector of estimated proportion of SNPs
belonging to each association pattern.\cr
\code{D_mat} \tab matrix of densities under each association pattern.\cr
\code{Gamma} \tab correlation matrix.\cr
\code{chi_mix} \tab matrix of \eqn{-2}log(\eqn{P})-values.\cr
\code{A} \tab vector of scaling factors under the alternative distributions.\cr
\code{df_alt} \tab vector of degrees of freedom approximated for the alternative distributions.\cr
}

The main element of interest for inference is the posterior probabilities matrix, \code{post_prob}.
The estimated proportion of observations belonging to each association pattern, \code{pis}, may
also be of interest. Note that the \code{pis} may not sum to 1 since they are estimated from Primo
run on a larger number of studies.
The remaining elements are returned primarily for use by other functions --
such as those conducting conditional association analysis.
}
\description{
For each SNP, estimates the posterior probability for each configuration.
Uses parameters estimated by previous runs of Primo to estimate probabilities
for SNPs missing in one or more studies. \eqn{P}-values from non-missing studies
are used as input. Utilizes parallel computing, when available.
}
