% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{Primo_ModT}
\alias{Primo_ModT}
\title{Estimate posterior probabilities of association patterns, using moderated t-statistics.}
\usage{
Primo_ModT(Tstat_mod, mdfs, V_mat, Gamma, tol = 0.001, par_size = 1)
}
\arguments{
\item{Tstat_mod}{matrix of moderated t-statistics.}

\item{mdfs}{matrix of moderated degrees of freedom.}

\item{Gamma}{correlation matrix.}

\item{tol}{numerical value; specifies tolerance threshold for convergence.}

\item{par_size}{numerical value; specifies the number of CPUs/cores/processors for
parallel computing (1 for sequential processing).}

\item{V}{matrix of scaling factors.}
}
\value{
A list with the following elements:
\tabular{ll}{
\code{post_prob} \tab matrix of posterior probabilities
(rows are SNPs; columns are association patterns)\cr
\code{pis} \tab vector of estimated proportion of SNPs
belonging to each association pattern\cr
\code{D_mat} \tab matrix of densities under each association pattern\cr
\code{Gamma} \tab correlation matrix\cr
\code{Tstat_mod} \tab matrix of moderated t-statistics\cr
\code{V_mat} \tab matrix of scaling factors under the alternative distribution\cr
\code{mdf_sd_mat} \tab matrix of standard deviation adjustment according to
 moderated degrees of freedom: df/(df-2)\cr
}
}
\description{
This version of the function uses moderated \eqn{t}-statistics and parameters
previously calculated under the limma framework. It is useful for cases where
the same statistic from one study (e.g. gene-SNP pair) may be mapped to
multiple statistics from another study (e.g. multiple gene-CpG pairs).
For each SNP, estimates the posterior probability for each configuration.
Utilizes parallel computing, when available.
}
\details{
The following are additional details describing the input arguments
 (for \eqn{m} SNPs/observations measured in \eqn{d} studies):
\tabular{ll}{
\code{Tstat_mod} \tab  \eqn{m} x \eqn{d} matrix\cr
\code{mdfs} \tab \eqn{m} x \eqn{d} matrix\cr
\code{V} \tab \eqn{m} x \eqn{d} matrix\cr
\code{Gamma} \tab  \eqn{d} x \eqn{d} matrix.\cr
}
}
