% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{Primo_tstat}
\alias{Primo_tstat}
\title{Estimate posterior probabilities of association patterns, using t-statistics.}
\usage{
Primo_tstat(betas, sds, dfs, alt_props, mafs = NULL, Gamma = NULL,
  tol = 0.001, par_size = 1)
}
\arguments{
\item{betas}{matrix of coefficient estimates.}

\item{sds}{matrix of standard errors (for coefficient estimates).}

\item{dfs}{vector or matrix of degrees of freedom.}

\item{alt_props}{vector of the proportions of test-statistics used in estimating
alternative densities.}

\item{mafs}{vector or matrix of minor allele frequencies (MAFs).}

\item{Gamma}{correlation matrix.}

\item{tol}{numerical value; specifies tolerance threshold for convergence.}

\item{par_size}{numerical value; specifies the number of CPUs/cores/processors for
parallel computing (1 for sequential processing).}
}
\value{
A list with the following elements:
\tabular{ll}{
\code{post_prob} \tab matrix of posterior probabilities
(rows are SNPs; columns are association patterns)\cr
\code{pis} \tab vector of estimated proportion of SNPs
belonging to each association pattern\cr
\code{D_mat} \tab matrix of densities under each association pattern\cr
\code{Tstat_mod} \tab matrix of moderated t-statistics\cr
\code{V_mat} \tab matrix of scaling factors under the alternative distribution\cr
\code{mdf_sd_mat} \tab matrix of standard deviation adjustment according to
 moderated degrees of freedom: df/(df-2)\cr
\code{Gamma} \tab correlation matrix\cr
\code{alt_props} \tab vector of the proportions of test-statistics used in
estimating alternative densities\cr
\code{tol} \tab numerical value; the tolerance threshold used in determining convergence
}
}
\description{
For each SNP, estimates the posterior probability for each configuration.
Utilizes parallel computing, when available.
}
\details{
The following are additional details describing the input arguments
 (for \eqn{m} SNPs/observations measured in \eqn{d} studies):
\tabular{ll}{
\code{betas} \tab  \eqn{m} x \eqn{d} matrix\cr
\code{sds} \tab \eqn{m} x \eqn{d} matrix\cr
\code{dfs} \tab vector of length \eqn{d} or an \eqn{m} x \eqn{d} matrix\cr
\code{alt_props} \tab vector of length \eqn{d}\cr
\code{mafs} \tab vector of length \eqn{d} or an \eqn{m} x \eqn{d} matrix.\cr
 \tab If not specified, standard errors will not be adjusted for MAF.\cr
\code{Gamma} \tab  \eqn{d} x \eqn{d} matrix.\cr
 \tab If not specified, will be estimated using observations where all \eqn{|t| < 5}.\cr
}
}
