% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional.R
\name{run_conditional}
\alias{run_conditional}
\title{Set up conditional analysis for a specified variant by index.}
\usage{
run_conditional(Primo_obj, IDs, idx, leadsnps_region, snp_col = "SNP",
  pheno_cols, snp_info, LD_mat, LD_thresh = 1, dist_thresh = 0,
  pval_thresh = 1, suffices = 1:length(pheno_cols))
}
\arguments{
\item{Primo_obj}{list returned by running the \eqn{t}-statistic version
of Primo (i.e. \code{\link{Primo_tstat}})}

\item{IDs}{data.frame of the SNP and phenotype IDs corresponding to each row
of \code{Primo_obj}.}

\item{idx}{integer specifying the index of the genetic variant (e.g. row of \code{Tstat_mod})
on which one wants to perform conditional analysis.}

\item{leadsnps_region}{data.frame detailing the lead SNP of each phenotype
in each region of \code{Primo_obj} and associated \eqn{P}-values for the lead SNPs.
See Details.}

\item{snp_col}{string of the column name of SNPs/variants.}

\item{pheno_cols}{character vector of the column names of the phenotype ID columns.}

\item{snp_info}{data.frame reporting the chromosome and position of each SNP.
Columns should be: \code{SNP, CHR, POS}.}

\item{LD_mat}{matrix of LD coefficients (\eqn{r^{2}}{r^2}). Row and column names
should be SNP/variant names (e.g. in \code{snp_col}).}

\item{LD_thresh}{scalar corresponding to the LD coefficient (\eqn{r^{2}}{r^2})
threshold to be used for conditional analysis. Lead SNPs with \eqn{r^{2} <}{r^2 <}
\code{LD_thresh} with the \code{idx} variant will be conditioned on.
Default value (1) signifies no consideration of LD in conditional analyses.}

\item{dist_thresh}{scalar of the minimum number of base pairs away from the \code{idx} SNP
that a lead SNP must be in order to be conditioned on. Default value (0)
signifies no consideration of chromosomal distance in conditional analyses.}

\item{pval_thresh}{scalar of the \eqn{P}-value threshold a lead SNP must be below
with the phenotype for which it is lead SNP in order to be conditionaed on.
Default value (1) signifies no consideration of strength of effect in conditional analyses.}

\item{suffices}{character vector of the suffices corresponding to columns in
\code{leadsnps_region}. See Details.}
}
\value{
The integer corresponding to the association pattern
with the highest posterior probability following conditional analysis.
The value returned, \eqn{k}, corresponds to the \eqn{k}-th column of \code{pis}
from the Primo output, and the \eqn{k}-th row of the \eqn{Q} matrix produced
by \code{\link{make_qmat}}.
}
\description{
For a specified variant (passed by index), set-up and run conditional analysis.
The function uses a data.frame of lead SNPs to identify possible SNPs
for conditional analysis, and determines which SNPs
will be conditioned on based on specified criteria. Returns the association
pattern with the highest posterior probability after conditional analysis.
}
\details{
The following are additional details describing the input
\code{leadsnps_region}. For \eqn{J} phenotypes being analyzed, the first
\eqn{J} columns of \code{leadsnps_region} should specify phenotype IDs.
Examples include gene symbol, CpG site name, trait name for GWAS, etc. The following
\eqn{2*J} columns should hold the name and p-value of the lead SNP for each of the
\eqn{J} phenotypes. The column names should be of the form \code{leadSNP_x} and
\code{pvalue_x}, where \code{x} is a suffix corresponding to the phenotype.
}
