% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdr.R
\name{select_ppthresh}
\alias{select_ppthresh}
\title{Posterior Probability threshold to control FDR}
\usage{
select_ppthresh(efdr_res, fdr = 0.05, tol = 0.001, descending = T,
  max_pp = T)
}
\arguments{
\item{efdr_res}{vector of eFDR values by potential posterior probability thresholds}

\item{fdr}{numerical value; desired/specified false discovery rate}

\item{tol}{the tolerance level for FDR control (i.e. half of the convergence window)}

\item{descending}{logical; denotes whether posterior probability thresholds are in
descending order}

\item{max_pp}{logical; denotes whether to return the maximum or minimum posterior probability
threshold which meets the FDR criteria}
}
\value{
numerical value; posterior probability threshold if selected efdr was named
after probability value.
Otherwise, returns the index of the matching eFDR (with warning message).
}
\description{
Determine posterior probability threshold to use for controlling false discovery rate (FDR).
The function a vector of empirical FDR calculated for candidate posterior thresholds.
and returns a posterior probability threshold expected to control FDR at specified rate.
}
