% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdr.R
\name{tfdr}
\alias{tfdr}
\title{Calculate theoretical False Discovery Rate (tFDR)}
\usage{
tfdr(post_prob, pp_thresh)
}
\arguments{
\item{post_prob}{vector of estimated posterior probabilities}

\item{pp_thresh}{numerical value; the posterior probability threshold at which
to calculate the tFDR}
}
\value{
Theoretical FDR (numeric) at specified posterior probability threshold.
}
\description{
Calculate the theoretical false discovery rate (tFDR) at a specified
posterior probability threshold.
}
\details{
The function estimates the theoretical FDR at the specified posterior probability
threshold by using the mean of \code{(1-post_prob)} for probabilities exceeding that threshold.
}
