% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdr.R
\name{tfdr_thresh}
\alias{tfdr_thresh}
\title{Calculate threshold for theoretical False Discovery Rate (tFDR)}
\usage{
tfdr_thresh(post_prob, fdr = 0.1)
}
\arguments{
\item{post_prob}{vector of estimated posterior probabilities}

\item{fdr}{numerical value; the desired false discovery rate (FDR)}
}
\value{
Posterior probability threshold (numeric) to use in controlling the FDR.
}
\description{
Determine the posterior probability threshold at which to control the theoretical
False Discovery Rate (tFDR) given observed posterior probabilities.
}
\details{
The function estimates the theoretical false discovery rate at each observed
posterior probability (PP) by sorting the PP in decreasing order and calculating
\code{cumsum(1-PP)/(#obs >= PP)}. It observed selects the first PP for which this value
is greater than \code{fdr}.
}
