% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calc_Dmatrix}
\alias{calc_Dmatrix}
\title{Density matrix calculation}
\usage{
calc_Dmatrix(Q, D0, D1)
}
\arguments{
\item{Q}{matrix of configurations.}

\item{D0}{matrix of density estimates under the null distribution.}

\item{D1}{matrix of density estimates under the alternative distributions.}
}
\value{
A matrix estimating the conditional joint densities
under each association pattern for each observation (e.g SNP).
}
\description{
Calculate conditional joint densities for each configuration, given marginal
null and alternative densities, under the assumption of independence.
}
