% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdr.R
\name{calc_fdr}
\alias{calc_fdr}
\title{Calculate the empirical false discovery rate (FDR).}
\usage{
calc_fdr(post_prob, threshold)
}
\arguments{
\item{post_prob}{numeric vector of posterior probabilities.}

\item{threshold}{numeric value of the posterior probability threshold at which to
calculate the empirical FDR.}
}
\value{
A numeric value of the empirical false discovery rate (FDR).
}
\description{
Calculate the empirical false discovery rate (FDR) given
a vector of posterior probabilities and a specified threshold.
}
