% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdr.R
\name{calc_fdr_col}
\alias{calc_fdr_col}
\title{Calculate the empirical false discovery rate (FDR) for a matrix column.}
\usage{
calc_fdr_col(post_prob, threshold, col_name = NULL, col_idx = NULL)
}
\arguments{
\item{post_prob}{numeric matrix of posterior probabilities.}

\item{threshold}{numeric value of the posterior probability threshold at which to
calculate the empirical FDR.}

\item{col_name}{character string of the column name.}

\item{col_idx}{integer of the column index.}
}
\value{
A numeric value of the empirical false discovery rate (FDR).
}
\description{
Calculate the empirical false discovery rate (FDR) of a
column in a matrix of posterior probabilities given a specified threshold.
}
\details{
At least one of \code{col_name} or \code{col_idx} must be specified.
}
