% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{collapse_pp_num}
\alias{collapse_pp_num}
\title{Collapse posterior probabilities based on number of traits of association.}
\usage{
collapse_pp_num(post_prob, req_idx = NULL, prefix = "pp_ge_")
}
\arguments{
\item{post_prob}{matrix of posterior probabilities.}

\item{req_idx}{(optional) integer vector of trait(s) whose association is required.
If \code{NULL}, no traits will be required.}

\item{prefix}{character string denoting prefix of column names for the results matrix.}
}
\value{
A numeric matrix of combined posterior probabilities.
}
\description{
Combine the posterior probabilities of each individual association pattern
according to the number of traits in each pattern. Provides the posterior probability of
being associated with "at least x" number of traits under analysis. Can also require
association with one or more traits and combine over other traits.
}
