% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{em_iter}
\alias{em_iter}
\title{EM Iteration}
\usage{
em_iter(old_pi, Q, D_0, D_1)
}
\arguments{
\item{old_pi}{vector of configuration proportions, fit through maximization
(usualy 2^J for J data types)}

\item{Q}{matrix of configurations}

\item{D_0}{estimate for the null density function}

\item{D_1}{estimate for the alternative density function}
}
\value{
Returns a vector estimating the proportion of SNPs coming from each configuration.
}
\description{
Complete one iteration of the EM-algorithm (combines E-step and M-step).
Calculate posterior expectations given current estimate for \eqn{\pi} (the
proportion of SNPs coming from each configuration; E-Step).
Re-estimates the \eqn{\pi} vector that maximizes the posterior expectation function from the E-step (i.e. M-step).
}
