% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{em_iter}
\alias{em_iter}
\title{EM Iteration}
\usage{
em_iter(old_pi, Dmat)
}
\arguments{
\item{old_pi}{vector of configuration proportions, fit through maximization.}

\item{Dmat}{matrix of conditional joint densities under each configuration.}
}
\value{
A row vector estimating the proportion of observations coming from
each association pattern.
}
\description{
Complete one iteration of the EM-algorithm (combines E-step and M-step).
Calculate posterior expectations given current estimate for \eqn{\pi},the
proportion of observations (e.g. SNPs) coming from each association pattern
(i.e. E-Step). Re-estimate the \eqn{\pi} vector that maximizes the posterior
expectations (i.e. M-step).
}
