% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrative_func.R
\name{estimate_config}
\alias{estimate_config}
\title{Estimate Posterior Probabilities of Configurations}
\usage{
estimate_config(pvals = NULL, betas = NULL, sds = NULL, mafs = NULL,
  dfs = NULL, alt_props, tol = 0.001, par_size = 0, density_list = NULL)
}
\arguments{
\item{pvals}{matrix of p-values from test statistics.}

\item{betas}{matrix of coefficient estimates.}

\item{sds}{matrix of standard errors (for coefficient estimates).}

\item{mafs}{vector of minor allele frequencies (MAFs).}

\item{dfs}{vector of degrees of freedom.}

\item{alt_props}{vector of the proportions of test-statistics used in estimating
alternative densities.}

\item{tol}{numerical value; specifies tolerance threshold for convergence.}

\item{par_size}{numerical value; specifies the number of CPUs/cores/processors for
parallel computing (0 for sequential processing).}

\item{density_list}{(optional) list of densities estimated by \code{estimate_densities()}.
If provided by user, must follow the format specified in Details.}
}
\value{
A list with the following elements:
\tabular{ll}{
\code{post_prob} \tab matrix of posterior probabilities
(rows are SNPs; columns are configurations)\cr
\code{config_prop} \tab vector of estimated proportion of SNPs
belonging to each configuration\cr
\code{Tstat_m} \tab matrix of moderated t-statistics\cr
\code{D0} \tab matrix of densities calculated under the null distribution\cr
\code{D1} \tab matrix of densities calculated under the alternative distribution\cr
\code{alt_props} \tab vector of the proportions of test-statistics used in
estimating alternative densities\cr
\code{tol} \tab numerical value; the tolerance threshold used in determining convergence
}
}
\description{
For each SNP, estimates the posterior probability for each configuration.
Utilizes parallel computing, when available.
}
\details{
Either \code{pvals} are all four of \code{betas}, \code{sds}, \code{mafs} and \code{dfs}
must be specified (i.e. not \code{NULL}).

The following are additional details describing the input arguments:
\tabular{ll}{
\code{pvals} \tab  p-values should be an LxD matrix: L = number of SNPs ("Loci");
D = number data sources.\cr
\code{betas} \tab  beta coefficients should be an LxD matrix: L = number of SNPs ("Loci");
D = number data sources.\cr
\code{sds} \tab standard errors should be an LxD matrix: L = number of SNPs ("Loci");
D = number data sources.\cr
\code{mafs} \tab minor allele frequencies (MAF) should be vector with length matching
the number of SNPs/Loci\cr
 \tab (same order as rows in \code{betas} and \code{sds}).\cr
\code{dfs} \tab vector containing number of subjects in each dataset\cr
 \tab (same order as columns in \code{betas} and \code{sds})\cr
\code{alt_props} \tab proportion of test-statistics used to estimate
the alternative distribution for each data source\cr
 \tab (same order as columns in \code{betas} and \code{sds})\cr
\code{density_list} \tab list must be 3 elements, named \code{Tstat_m},\code{D0}, and \code{D1}
 -- outputs from \code{estimate_densities()}\cr
 \tab (each element is a matrix, with row matching observations of \code{betas} and \code{sds},\cr
 \tab and columns in the same order as columns in \code{betas} and \code{sds})
}
}
