% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density_estimation.R
\name{estimate_densities}
\alias{estimate_densities}
\title{Estimate parameters under the null and alternative densities}
\usage{
estimate_densities(betas = NULL, sds = NULL, df = NULL,
  pvals = NULL, alt_prop, mafs = NULL, use_tstats = TRUE)
}
\arguments{
\item{betas}{vector of coefficient estimates.}

\item{sds}{vector of standard errors (for coefficient estimates).}

\item{df}{scalar or vector of first degrees of freedom of the F-distribution.}

\item{pvals}{vector of \eqn{P}-values.}

\item{alt_prop}{proportion of test-statistics used in estimating alternative densities.}

\item{mafs}{vector of minor allele frequencies (MAFs).}

\item{use_tstats}{logical indicating whether to use moderated \eqn{t}-statistics (\code{TRUE})
or \eqn{P}-values (\code{FALSE}).}
}
\value{
See \code{\link{estimate_densities_modT}} or \code{\link{estimate_densities_pval}}.
}
\description{
Estimate the parameters of the null and alternative
density functions.
This function calls either \code{\link{estimate_densities_modT}} or
\code{\link{estimate_densities_pval}} for density estimation using
moderated \eqn{t}-statistics or \eqn{P}-values, respectively.
}
