% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density_estimation.R
\name{estimate_densities_modT}
\alias{estimate_densities_modT}
\title{Estimate Densities Under the Null and Alternative Densities}
\usage{
estimate_densities_modT(betas, sds, mafs = NULL, df, alt_prop)
}
\arguments{
\item{betas}{vector of coefficient estimates.}

\item{sds}{vector of standard errors (for coefficient estimates).}

\item{mafs}{vector of minor allele frequencies (MAFs) (optional).
If \code{NULL}, no adjustment will be made for MAF in variance calculations.}

\item{df}{first degrees of freedom of the F-distribution (usually, the number of subjects/observations).}

\item{alt_prop}{proportion of test-statistics used in estimating alternative densities.}
}
\value{
A list with the following elements:
\tabular{ll}{
\code{Tstat_mod} \tab matrix of moderated t-statistics\cr
\code{df_mod} \tab vector of the moderated degrees of freedom\cr
\code{scaler} \tab vector of the scaling factors for the moderated t-statistics under the alternative\cr
}
}
\description{
For each observation, estimate the density under the null and under the alternative
hypotheses using moderated t-statistics.
}
\details{
Following Smyth (2004), the function calculates moderated t-statistics using
conditional posterior means of the variance. The moderated t-statistics are used to estimate
the null density under a t-distribution with estimated degrees of freedom and
the alternative density under a scaled t-distribution.
}
