% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density_estimation.R
\name{estimate_densities_modT}
\alias{estimate_densities_modT}
\title{Estimate parameters under the null and alternative densities}
\usage{
estimate_densities_modT(betas, sds, mafs = NULL, df, alt_prop,
  N = NULL)
}
\arguments{
\item{betas}{vector of coefficient estimates.}

\item{sds}{vector of standard errors (for coefficient estimates).}

\item{mafs}{vector of minor allele frequencies (MAFs) (optional).
If \code{NULL}, no adjustment will be made for MAF in variance calculations.}

\item{df}{scalar or vector of first degrees of freedom of the F-distribution
(usually, \eqn{n-p-1}).}

\item{alt_prop}{proportion of test-statistics used in estimating the alternative density.}

\item{N}{number of subjects}
}
\value{
A list with the following elements:
\tabular{ll}{
\code{Tstat_mod} \tab matrix of moderated t-statistics\cr
\code{df_mod} \tab vector of the moderated degrees of freedom\cr
\code{scaler} \tab vector of the scaling factors for the moderated t-statistics under the alternative\cr
}
}
\description{
Estimate the parameters of the null and alternative
density functions using moderated \eqn{t}-statistics.
}
\details{
Following Smyth (2004), the function calculates moderated \eqn{t}-statistics using
conditional posterior means of the variance. The moderated \eqn{t}-statistics are used to estimate
the null density under a \eqn{t}-distribution and
the alternative density under a scaled \eqn{t}-distribution, each with moderated degrees of freedom.
}
