% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrative_func.R
\name{estimate_densities_pval}
\alias{estimate_densities_pval}
\title{Estimate Densities Under the Null and Alternative Densities}
\usage{
estimate_densities_pval(pvals, alt_proportion)
}
\arguments{
\item{pvals}{vector of coefficient estimates.}

\item{alt_proportion}{proportion of pvalues used in estimating alternative densities.
Note that the alternative density estimation may fail
if the value is too small.}
}
\value{
A list with the following elements:
\tabular{ll}{
\code{D0} \tab matrix of densities calculated under the null distribution\cr
\code{D1} \tab matrix of densities calculated under the alternative distribution\cr
}
}
\description{
For each observation, estimate the density under the null and under the alternative
hypotheses using p-values.
}
\details{
The function estimates densities under the null and under the alternative
given a vector of p-values from test statistics and the marginal probability of
coming from the alternative distribution. Under the null hypothesis,
\code{-2*log()} transformed p-values follow a chi-square distribution with 2 degrees of freedom.
Under the alternative, the function assumes that the \code{-2*log()} transformed p-values
follow a scaled chi-square distribution with unknown scale parameter and
unknown degrees of freedom. Those two parameters are solved by using the first and
second moments of the transformed p-values and the known proportion of alternatives
(specified in \code{alt_proportion}). The function solves a theoretical formula of the first and
second moments as functions of the scale parameter and degrees of freedom.
}
