% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{m_step}
\alias{m_step}
\title{M-step}
\usage{
m_step(old_B)
}
\arguments{
\item{old_B}{matrix of posterior expectations/probabilities (e.g. from an E-step).}
}
\value{
A row vector estimating the proportion of observations coming from each association pattern.
}
\description{
Estimates the \eqn{\pi} vector that maximizes the posterior expectation function from the E-step.
}
