% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrative_func.R
\name{make_qmat}
\alias{make_qmat}
\title{Make Configuration Matrix}
\usage{
make_qmat(grp, name.cols = F)
}
\arguments{
\item{grp}{vector of group identifiers (common to use seq(1,N), for N groups)}

\item{name.cols}{logical; whether or not to name columns using group identifiers}
}
\value{
\code{make_qmat} returns a matrix with all binary combinations of the groups
}
\description{
Make configuration matrix (Q) that indicates null/alternative density patterns.
}
