% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density_estimation.R
\name{make_qmat}
\alias{make_qmat}
\title{Make Configuration Matrix}
\usage{
make_qmat(grp, name_cols = FALSE)
}
\arguments{
\item{grp}{vector of group identifiers (common to use \code{seq(1,d)}, for \eqn{d} groups).}

\item{name_cols}{logical indicating whether to name columns using group identifiers.}
}
\value{
A matrix with all binary combinations of the groups. The value in row \eqn{k}
and column \eqn{j} denotes whether statistics from study \eqn{j} come from
the null (\eqn{0}) or alternative (\eqn{1}) distribution
under the \eqn{k}-th association pattern.
}
\description{
Make configuration matrix (\eqn{Q}) that indicates null/alternative densities
in the association patterns.
}
