% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permute_func.R
\name{permute_setup}
\alias{permute_setup}
\title{Estimate Posterior Probabilities for Permuted Datasets}
\usage{
permute_setup(betas, sds, mafs, dfs, true_res, par_size = 0,
  perm_par_size = 0)
}
\arguments{
\item{betas}{matrix of coefficient estimates.}

\item{sds}{matrix of standard errors (for coefficient estimates).}

\item{mafs}{vector of minor allele frequencies (MAFs).}

\item{dfs}{vector of degrees of freedom.}

\item{true_res}{a list; results returned by \code{\link{estimate_config}} when
run with true data}

\item{par_size}{numerical value; specifies the number of CPUs/cores/processors for
parallel computing (0 for sequential processing).}
}
\value{
A list of lists, where each list holds results from a single run
of \code{\link{estimate_config}}.Each run represents results using one permuted dataset.
See \code{\link{estimate_config}} for details of each list.
}
\description{
Wrapper for \code{\link{permute_integ}} that ensures the
same parameters are used for permuted datasets as was used
when calculating posterior probabilities for the true data.
Permutes several datasets (one column permuted in each)
and estimates the posterior probability for each configuration
for each SNP under each permuted dataset.
Utilizes parallel computing, when available.
}
\details{
See documentation for \code{\link{estimate_config}} for additional details
regarding the input arguments.
}
