% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Primo.R
\docType{package}
\name{Primo-package}
\alias{Primo-package}
\alias{_PACKAGE}
\title{Primo: Package in R for Integrating Multi-Omics data}
\description{
Implements methods to integrate multi-omics data to detect joint associations 
              and pleiotropy effects across more than 3 QTL and GWAS studies. 
              It can also be used more broadly to infer multi-way associations 
              using omics data from more than 2 data types or sources, beyond QTL and GWAS.
}
\details{
To estimate the posterior probabilities of belonging to each association pattern,
either user the wrapper function \code{\link{Primo}}, or directly call
\code{\link{Primo_tstat}} (for t-statistics) or \code{\link{Primo_pval}} (for p-values).
}
\author{
\strong{Maintainer}: Kevin J. Gleason \email{kevin.j.gleason@gmail.com}

Authors:
\itemize{
  \item Fan Yang \email{FAN.3.YANG@ucdenver.edu}
  \item Lin S. Chen \email{lchen@health.bsd.uchicago.edu}
}

}
