% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{Primo_modT}
\alias{Primo_modT}
\title{Estimate posterior probabilities of association patterns, using moderated t-statistics.}
\usage{
Primo_modT(Tstat_mod, mdfs, V_mat, Gamma, tol = 0.001, par_size = 1)
}
\arguments{
\item{Tstat_mod}{matrix of moderated t-statistics.}

\item{mdfs}{matrix of moderated degrees of freedom.}

\item{V_mat}{matrix of scaling factors.}

\item{Gamma}{correlation matrix.}

\item{tol}{numeric value specifying the tolerance threshold for convergence.}

\item{par_size}{numeric value specifying the number of workers for
parallel computing (1 for sequential processing).}
}
\value{
A list with the following elements:
\tabular{ll}{
\code{post_prob} \tab matrix of posterior probabilities
(each column corresponds to an association pattern).\cr
\code{pis} \tab vector of estimated proportion of observations
belonging to each association pattern.\cr
\code{D_mat} \tab matrix of densities under each association pattern.\cr
\code{Gamma} \tab correlation matrix.\cr
\code{Tstat_mod} \tab matrix of moderated t-statistics.\cr
\code{V_mat} \tab matrix of scaling factors under the alternative distribution.\cr
\code{mdf_sd_mat} \tab matrix of standard deviation adjustment according to
 moderated degrees of freedom: df/(df-2).\cr
}

The main element of interest for inference is the posterior probabilities matrix, \code{post_prob}.
The estimated proportion of observations belonging to each association pattern, \code{pis}, may
also be of interest. The remaining elements are returned primarily for use by other functions.
}
\description{
This version of the main \code{Primo} function uses moderated \eqn{t}-statistics
and parameters previously calculated under the limma framework
(i.e. using \code{\link{estimate_densities_modT}}).
It is useful for cases where the same statistic from one study
(e.g. gene-SNP pair) may be mapped to multiple statistics from
another study (e.g. multiple gene-CpG pairings). For each observation
(e.g. SNP), it estimates the posterior probability for each association pattern.
Utilizes parallel computing, when available.
}
\details{
The following are additional details describing the input arguments
 (for \eqn{m} SNPs/observations measured in \eqn{d} studies):
\tabular{ll}{
\code{Tstat_mod} \tab  \eqn{m} x \eqn{d} matrix.\cr
\code{mdfs} \tab \eqn{m} x \eqn{d} matrix.\cr
\code{V_mat} \tab \eqn{m} x \eqn{d} matrix.\cr
\code{Gamma} \tab  \eqn{d} x \eqn{d} matrix.\cr
}
}
