% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{Primo_pval}
\alias{Primo_pval}
\title{Estimate posterior probabilities of association patterns, using P-values.}
\usage{
Primo_pval(pvals, alt_props, Gamma = NULL, tol = 0.001)
}
\arguments{
\item{pvals}{matrix of \eqn{P}-values from test statistics.}

\item{alt_props}{vector of the proportions of test-statistics used in estimating
alternative densities.}

\item{Gamma}{correlation matrix.}

\item{tol}{numerical value; specifies tolerance threshold for convergence.}

\item{par_size}{numerical value; specifies the number of CPUs/cores/processors for
parallel computing (1 for sequential processing).}
}
\value{
A list with the following elements:
\tabular{ll}{
\code{post_prob} \tab matrix of posterior probabilities
(rows are SNPs; columns are association patterns)\cr
\code{pis} \tab vector of estimated proportion of SNPs
belonging to each association pattern\cr
\code{D_mat} \tab matrix of densities under each association pattern\cr
\code{Gamma} \tab correlation matrix\cr
\code{chi_mix} \tab matrix of \eqn{-2}log(\eqn{P})-values \cr
\code{A} \tab vector of scaling factors under the alternative distribution\cr
\code{df_alt} \tab vector of degrees of freedom approximated for the alternative distributions\cr
}
}
\description{
For each SNP, estimates the posterior probability for each configuration.
Utilizes parallel computing, when available.
}
\details{
The following are additional details describing the input arguments
 (for \eqn{m} SNPs/observations measured in \eqn{d} studies):
\tabular{ll}{
\code{pvals} \tab  \eqn{m} x \eqn{d} matrix\cr
\code{alt_props} \tab vector of length \eqn{d}\cr
\code{Gamma} \tab  \eqn{d} x \eqn{d} matrix.
 If not specified, will be estimated using observations where all \eqn{p < 5.7e-7}.\cr
}
}
