% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional.R
\name{run_conditional}
\alias{run_conditional}
\title{Set up conditional analysis for specified variant(s) by index.}
\usage{
run_conditional(Primo_obj, IDs, idx, leadsnps_region, snp_col = "SNP",
  pheno_cols, snp_info, LD_mat, LD_thresh = 1, dist_thresh = 0,
  pval_thresh = 1, suffices = 1:length(pheno_cols))
}
\arguments{
\item{Primo_obj}{list returned by running the \eqn{t}-statistic version
of Primo (i.e. \code{\link{Primo_tstat}})}

\item{IDs}{data.table of the SNP and phenotype IDs corresponding to each row
of the Primo results stored in \code{Primo_obj}.}

\item{idx}{integer of the index of the genetic variant (e.g. row of \code{Tstat_mod})
on which one wants to perform fine-mapping}

\item{leadsnps_region}{data.table that stores the lead SNP of each phenotype
in each region of the Primo results. Also includes p-values for the lead SNPs.
See Details for format.}

\item{snp_col}{string of the column name of SNPs/variants.}

\item{pheno_cols}{character vector of the column names of the phenotype ID columns.}

\item{snp_info}{data.table reporting the chromosome and position of each SNP.
Columns must include: \code{SNP, CHR, POS}.}

\item{LD_mat}{matrix of LD coefficients (\eqn{r^{2}}{r^2}). Row and column names
should be SNP/variant names (e.g. in \code{snp_col}).}

\item{LD_thresh}{scalar corresponding to the LD coefficient (\eqn{r^{2}}{r^2})
threshold to be used for conditional analysis. Lead SNPs with \eqn{r^{2} <}{r^2 <}
\code{LD_thresh} with the \code{idx} variant will be conditioned on.
Default value (1) signifies no consideration of LD in conditional analyses.}

\item{dist_thresh}{scalar of the minimum number of base pairs away from the \code{idx} SNP
that a lead SNP must be in order to be conditioned on. Default value (0)
signifies no consideration of chromosomal distance in conditional analyses.}

\item{pval_thresh}{scalar of the \eqn{P}-value threshold a lead SNP must be below
with the phenotype for which it is lead SNP in order to be conditionaed on.
Default value (1) signifies no consideration of strength of effect in conditional analyses.}

\item{suffices}{character vector of the suffices corresponding to columns in
\code{leadsnps_region}. See Details.}
}
\value{
An integer vector corresponding to the association pattern
with the highest posterior probabilities for each SNP variant
represented by \code{idx}, following conditional analysis.
Each value returned, \eqn{k}, corresponds to the \eqn{k}-th column of \code{pis}
from the Primo output, and the \eqn{k}-th row of the \eqn{Q} matrix produced
by \code{\link{make_qmat}}.
}
\description{
For specified variant(s) (passed by index), set-up and run conditional analysis.
The function uses a data.table of lead SNPs to identify possible SNPs
for conditional analysis, and determines which SNPs
will be conditioned on based on specified criteria. Returns the association
pattern with the highest posterior probability for each specified variant
after conditional analysis.
}
