% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdr.R
\name{tfdr_thresh}
\alias{tfdr_thresh}
\title{Calculate threshold for theoretical False Discovery Rate (tFDR)}
\usage{
tfdr_thresh(post_prob, fdr)
}
\arguments{
\item{post_prob}{vector of estimated posterior probabilities}

\item{fdr}{numerical value; the desired false discovery rate (FDR)}
}
\value{
Posterior probability threshold (numeric) to use in controlling the FDR.
}
\description{
Determine the posterior probability threshold at which to control the
false discovery rate (FDR) by estimating the theoretical FDR given the observed
posterior probabilities.
}
\details{
The function calculates the theoretical FDR at each observed posterior probability
by averaging \code{(1-post_prob)} over observations at least as large as that probability.
It returns the value of the first observation in \code{post_prob} for which the cumulative
average is greater than \code{fdr}.
}
